#39: moved types and enums to contracts
----
Create "contracts" package
https://github.com/monumentjs/core/issues/39
