+++ packages/collections/specs/unit/multivaluemap/mutable/MultiValueMap.spec.ts
@@ -1,5 +1,4 @@
-import { KeyValuePair } from '../../../../index';
-import { MultiValueMap } from '@monument/contracts';
+import { KeyValuePair, MultiValueMap } from '@monument/contracts';
 import { assertLengthAndIsEmpty, testReadOnlyMultiValueMap } from '../readonly/ReadOnlyMultiValueMap.spec';
 
 export function testMultiValueMap(create: <K, V>(items?: Iterable<KeyValuePair<K, V>>) => MultiValueMap<K, V>) {
