+++ packages/collections/specs/unit/multivaluemap/readonly/ReadOnlyMultiValueMap.spec.ts
@@ -1,5 +1,8 @@
-import { KeyValuePair } from '../../../../index';
-import { ReadOnlyMap, ReadOnlyMultiValueMap } from '@monument/contracts';
+import {
+  KeyValuePair,
+  ReadOnlyMap,
+  ReadOnlyMultiValueMap
+} from '@monument/contracts';
 
 export function assertLengthAndIsEmpty(map: ReadOnlyMultiValueMap<any, any>, expectedLength: number, expectedValuesCount: number): void {
   expect(map.length).toBe(expectedLength);
