+++ packages/contracts/src/collections/maps/ReadOnlyMultiValueMap.ts
@@ -1,12 +1,11 @@
 import { ToArray } from '../../data/ToArray';
 import { Sequence } from '../../data/Sequence';
-import { KeyValuePair } from '@monument/collections';
+import { KeyValuePair } from '../../types/KeyValuePair';
 import { ReadOnlyMap } from './ReadOnlyMap';
 import { Equatable } from '../../data/Equatable';
 import { SupplyFunction } from '../../types/SupplyFunction';
 
 /**
- * @todo need to change implementations to abstractions
  * @author Alex Chugaev
  * @since 0.0.1
  * @readonly
