+++ packages/contracts/src/date/DayOfWeek.ts
