+++ packages/contracts/src/types/function/AggregateFunction.ts
