+++ packages/contracts/src/types/function/IteratorFunction.ts
