+++ packages/contracts/src/types/function/MapIteratorFunction.ts
