+++ packages/date/index.ts
@@ -4,7 +4,6 @@ export * from './src/Date';
 export * from './src/DateTime';
 export * from './src/DateTimeLayout';
 export * from './src/DayOfMonth';
-export * from './src/DayOfWeek';
 export * from './src/DayOfYear';
 export * from './src/Duration';
 export * from './src/Hours';
@@ -12,7 +11,6 @@ export * from './src/InvariantDateTimeFormat';
 export * from './src/Milliseconds';
 export * from './src/Minutes';
 export * from './src/Month';
-export * from './src/NativeDate';
 export * from './src/Seconds';
 export * from './src/Time';
 
@@ -22,4 +20,4 @@ export * from './src/TimeZoneOffset';
 export * from './src/Year';
 
 // backward compatibility
-export { DateFields, DateTimeFields, DateTimeFormat, TimeFields } from '@monument/contracts';
+export { DateFields, DateTimeFields, DateTimeFormat, DayOfWeek, NativeDate, TimeFields } from '@monument/contracts';
