+++ packages/logger/index.ts
@@ -1,6 +1,5 @@
 // Export public API
 
-export * from './src/core/Level';
 export * from './src/core/LogAction';
 export * from './src/core/LogEvent';
 export * from './src/core/Logger';
@@ -19,4 +18,4 @@ export * from './src/transport/console/ConsoleTransport';
 export * from './src/transport/testing/LogEventHistory';
 
 // backward compatibility
-export { LoggingConfiguration, Filter, Transport, ConsoleLayout  } from '@monument/contracts';
+export { LoggingConfiguration, Filter, Level, Transport, ConsoleLayout  } from '@monument/contracts';
