+++ packages/store/index.ts
@@ -1,13 +1,13 @@
 // Export public API
 
 export * from './src/Actions';
+export * from './src/EffectResult';
+export * from './src/EffectSource';
 export * from './src/Catch';
 export * from './src/CatchDecorator';
 export * from './src/Effect';
 export * from './src/EffectDecorator';
 export * from './src/EffectMediator';
-export * from './src/EffectResult';
-export * from './src/EffectSource';
 export * from './src/Errors';
 export * from './src/ErrorMediator';
 export * from './src/Reaction';
