+++ packages/version/index.ts
@@ -1,9 +1,8 @@
 // Export public API
 
-export * from './src/ReleaseStatus';
 export * from './src/Version';
 export * from './src/VersionBuilder';
 
 // backward compatibility layer
 export { VersionFormatException } from '@monument/exceptions';
-export { VersionComponents } from '@monument/contracts';
+export { VersionComponents, ReleaseStatus } from '@monument/contracts';
